can-bridge: Refactor and organize into lib and main
